Linear Algebraic Equation System Approximate solution of the system of linear algebraic equations using single iteration and Seidel iteration methods.
Text Files
Text files are placed in the Texts directory.

input.txt is the file containing the default input matrix. Additionally, in the directory, there are log files storing the results of each iteration step.

Source Code Files
The program consists of 1 main source file and 4 auxiliary files placed in the Packages directory.

main.cpp
Main program file.

Console.cpp
File containing functions used to customize the console and draw the Menu window.

Each character on the console screen corresponds to a unit, and the maximum number of characters that can be displayed corresponds to the console size.
The program will have the following console window size:

Width: 185 units
Height: 50 units
Additionally, the maximum size of the program is 200x50.
The input area will start from the 36th row and have the same width as the maximum width.
The output area will start from the 4th column and have a height from the 4th row to the 46th row.
The program will disable the ability to zoom in, zoom out, and resize the window to ensure the best display results.
The results to be displayed in the output area will be converted into a vector with the data type being a string, where each element corresponds to the content of a row. This vector will be passed to the DrawOutputBox(vector<string>*) function to print out the output area. If the number of rows to be printed is too large, the program will display an error.

The program uses ANSI code to change colors and move the cursor on the application screen. The character set (code page) used is Unicode (utf-8) with ID 65001, which supports displaying Vietnamese.

Functions.cpp
File containing support functions for the algorithm.

The program allows the input system of equations to have a maximum of 10 unknowns.

The file contains functions for reading, writing files, calculating error for methods, processing matrices before printing to the output area.

By default, the program will display 4 decimal places, and this number can be changed up to a maximum of 6 decimal places.

MathFuncs.cpp
File containing some matrix calculation functions.

ColorCode.h
File containing definitions of color codes used in the program.